pyproject.toml
[project]
name = "asterics-grid-helper"
version = "0.1.0"
description = "Helper services for AsTeRICS Grid, providing text-to-speech and Signal messaging automation"
authors = [
    { name = "Will Wade", email = "willwade@gmail.com" }
]
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "flask-restx>=1.3.0",
    "py3-tts-wrapper[avsynth]; sys_platform == 'darwin'",
    "py3-tts-wrapper[sapi]; sys_platform == 'win32'",
    "py3-tts-wrapper[espeak]; sys_platform == 'linux'",
    "py3-tts-wrapper[sherpaonnx,espeak,elevenlabs,playht,microsoft,polly,watson,googletrans,witai,controlaudio]",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { file = "LICENSE" }

[tool.setuptools]
packages = ["speech"]

[project.optional-dependencies]
test = [
    "requests",
    "pytest",
    "pytest-cov",
    "pyinstaller",
    "hydra-core",
    "omegaconf",
    "pytest-mock>=3.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]

[tool.pytest.ini_options]
testpaths = ["speech"]
python_files = ["test_*.py"]
addopts = "-v --cov=speech --cov-report=term-missing" 